requirements.txt
# Fenrir - Network Security Scanner
# Python dependencies

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.0

# Database
sqlalchemy==2.0.35
aiosqlite==0.20.0

# HTTP client (for OpenRouter + NVD API)
httpx==0.27.2
aiohttp==3.10.5

# Environment
python-dotenv==1.0.1

# Nmap wrapper
python-nmap==0.7.1

# Data validation
pydantic==2.9.2
pydantic-settings==2.5.2

# Report generation
markdown==3.7
weasyprint==62.3

# Utilities
ipaddress==1.0.23
rich==13.8.1
